+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Buildings are now filtered on map according to selected level
+
 
 ## 1.0.2 - 2019-09-26